chore: migrate expiration modal
