+++ packages/apps/human-app/frontend/src/modules/auth/components/index.ts
@@ -1,0 +1,1 @@
+export * from './expiration-modal';
