+++ packages/apps/human-app/frontend/src/modules/auth/context/auth-context.tsx
@@ -1,5 +1,4 @@
 /* eslint-disable camelcase */
-// web2-auth.tsx
 import { z } from 'zod';
 import { createAuthProvider } from '@/shared/contexts/generic-auth-context';
