+++ packages/apps/human-app/frontend/src/modules/auth/hooks/use-expiration-modal.tsx
@@ -1,0 +1,15 @@
+import { useModal } from '@/shared/contexts/modal-context';
+import { ExpirationModal } from '@/modules/auth/components';
+
+export function useExpirationModal() {
+  const { openModal } = useModal();
+
+  return {
+    open: () => {
+      openModal({
+        content: <ExpirationModal />,
+        showCloseButton: false,
+      });
+    },
+  };
+}
