+++ packages/apps/human-app/frontend/src/modules/auth/index.ts
@@ -1,0 +1,5 @@
+export {
+  useAuth,
+  useAuthenticatedUser,
+  useExpirationModal,
+} from './hooks/index';
