+++ packages/apps/human-app/frontend/src/shared/components/ui/modal/modal-content.tsx
@@ -1,11 +1,9 @@
 import type { ComponentType } from 'react';
 import { WalletConnectModal } from '@/modules/auth-web3/components/wallet-connect-modal';
-import { ExpirationModal } from '@/modules/auth/components/expiration-modal';
 import { ModalType } from './modal.store';
 
 const MODAL_COMPONENTS_MAP: Record<ModalType, ComponentType> = {
   [ModalType.WALLET_CONNECT]: WalletConnectModal,
-  [ModalType.EXPIRATION_MODAL]: ExpirationModal,
 };
 
 interface ModalContent {
