+++ packages/apps/human-app/frontend/src/shared/components/ui/modal/modal.store.ts
@@ -4,7 +4,6 @@ import type { DialogProps as DialogMuiProps } from '@mui/material/Dialog';
 
 export enum ModalType {
   WALLET_CONNECT = 'WALLET_CONNECT',
-  EXPIRATION_MODAL = 'EXPIRATION_MODAL',
 }
 interface ModalState {
   isModalOpen: boolean;
